// src/google/longrunning/operations_pb.d.ts
export class Operation extends jspb.Message { 
    getName(): string;
    setName(value: string): Operation;

    hasMetadata(): boolean;
    clearMetadata(): void;
    getMetadata(): google_protobuf_any_pb.Any | undefined;
    setMetadata(value?: google_protobuf_any_pb.Any): Operation;
    getDone(): boolean;
    setDone(value: boolean): Operation;

    hasError(): boolean;
    clearError(): void;
    getError(): google_rpc_status_pb.Status | undefined;
    setError(value?: google_rpc_status_pb.Status): Operation;

    hasResponse(): boolean;
    clearResponse(): void;
    getResponse(): google_protobuf_any_pb.Any | undefined;
    setResponse(value?: google_protobuf_any_pb.Any): Operation;

    getResultCase(): Operation.ResultCase;

    serializeBinary(): Uint8Array;
    toObject(includeInstance?: boolean): Operation.AsObject;
    static toObject(includeInstance: boolean, msg: Operation): Operation.AsObject;
    static extensions: {[key: number]: jspb.ExtensionFieldInfo<jspb.Message>};
    static extensionsBinary: {[key: number]: jspb.ExtensionFieldBinaryInfo<jspb.Message>};
    static serializeBinaryToWriter(message: Operation, writer: jspb.BinaryWriter): void;
    static deserializeBinary(bytes: Uint8Array): Operation;
    static deserializeBinaryFromReader(message: Operation, reader: jspb.BinaryReader): Operation;
}

export namespace Operation {
    export type AsObject = {
        name: string,
        metadata?: google_protobuf_any_pb.Any.AsObject,
        done: boolean,
        error?: google_rpc_status_pb.Status.AsObject,
        response?: google_protobuf_any_pb.Any.AsObject,
    }

    export enum ResultCase {
        RESULT_NOT_SET = 0,
        ERROR = 4,
        RESPONSE = 5,
    }

}

export class GetOperationRequest extends jspb.Message { 
    getName(): string;
    setName(value: string): GetOperationRequest;

    serializeBinary(): Uint8Array;
    toObject(includeInstance?: boolean): GetOperationRequest.AsObject;
    static toObject(includeInstance: boolean, msg: GetOperationRequest): GetOperationRequest.AsObject;
    static extensions: {[key: number]: jspb.ExtensionFieldInfo<jspb.Message>};
    static extensionsBinary: {[key: number]: jspb.ExtensionFieldBinaryInfo<jspb.Message>};
    static serializeBinaryToWriter(message: GetOperationRequest, writer: jspb.BinaryWriter): void;
    static deserializeBinary(bytes: Uint8Array): GetOperationRequest;
    static deserializeBinaryFromReader(message: GetOperationRequest, reader: jspb.BinaryReader): GetOperationRequest;
}

export namespace GetOperationRequest {
    export type AsObject = {
        name: string,
    }
}